Add +x to magneto debug, add check for required file

+++ add-magneto-exclude.sh
@@ -1,7 +1,13 @@
-echo "public/app/code/community/Magneto/" | tee -a .git/info/exclude > /dev/null
-echo "public/app/design/adminhtml/base/default/layout/debug.xml" | tee -a .git/info/exclude > /dev/null
-echo "public/app/design/adminhtml/base/default/template/debug" | tee -a .git/info/exclude > /dev/null
-echo "public/app/design/frontend/base/default/layout/debug.xml" | tee -a .git/info/exclude > /dev/null
-echo "public/app/design/frontend/base/default/template/debug" | tee -a .git/info/exclude > /dev/null
-echo "public/app/etc/modules/Magneto_Debug.xml" | tee -a .git/info/exclude > /dev/null
-echo "public/skin/frontend/base/default/debug" | tee -a .git/info/exclude > /dev/null
+if [ -f .git/info/exclude ]; then
+    echo "# Magneto Debug" | tee -a .git/info/exclude > /dev/null
+    echo "public/app/code/community/Magneto/" | tee -a .git/info/exclude > /dev/null
+    echo "public/app/design/adminhtml/base/default/layout/debug.xml" | tee -a .git/info/exclude > /dev/null
+    echo "public/app/design/adminhtml/base/default/template/debug" | tee -a .git/info/exclude > /dev/null
+    echo "public/app/design/frontend/base/default/layout/debug.xml" | tee -a .git/info/exclude > /dev/null
+    echo "public/app/design/frontend/base/default/template/debug" | tee -a .git/info/exclude > /dev/null
+    echo "public/app/etc/modules/Magneto_Debug.xml" | tee -a .git/info/exclude > /dev/null
+    echo "public/skin/frontend/base/default/debug" | tee -a .git/info/exclude > /dev/null
+    echo "Done!"
+else
+    echo "You need to run this script from a git repository root!"
+fi
